Revert (接口定义): 接口定义优化需求回退
This reverts commit 2ff5769c
