+++ frontend/src/business/components/api/definition/components/case/ApiCaseItem.vue
@@ -345,7 +345,6 @@ export default {
         this.currentApi.request = item.request;
         this.currentApi.request.changeId = getUUID();
       }
-      this.$emit("setSelectedCaseId",item.id);
     },
     changePriority(row) {
       if (row.id) {
@@ -465,7 +464,6 @@ export default {
           if (!hideAlert) {
             this.$emit('refresh');
           }
-          this.$emit("saveCaseCallback");
         }
       }, (error) => {
         this.isSave = false;
