+++ frontend/src/business/components/api/definition/components/mock/MockTab.vue
@@ -13,7 +13,7 @@
 
       <ms-table
         v-loading="result.loading"
-        :data="filterMockData()"
+        :data="mockConfigData.mockExpectConfigList.filter(data=>!tableSearch || data.name.toLowerCase().includes(tableSearch.toLowerCase()))"
         :operators="operators"
         :page-size="pageSize"
         :showSelectAll="false"
@@ -72,10 +72,9 @@
         </ms-table-column>
       </ms-table>
     </div>
-    <mock-edit-drawer :is-tcp="isTcp" :api-id="apiId"
+    <mock-edit-drawer :is-tcp="isTcp" :api-id="this.baseMockConfigData.mockConfig.apiId"
                       :api-params="apiParams"
                       @refreshMockInfo="refreshMockInfo"
-                      v-if="mockConfigData && mockConfigData.mockConfig && mockConfigData.mockConfig.id"
                       :mock-config-id="mockConfigData.mockConfig.id" ref="mockEditDrawer"/>
   </div>
 </template>
@@ -96,8 +95,7 @@ export default {
   },
   props: {
     baseMockConfigData: {},
-    versionName: String,
-    apiId: String,
+    versionName:String,
     isTcp: {
       type: Boolean,
       default: false,
@@ -153,14 +151,6 @@ export default {
     },
   },
   methods: {
-    filterMockData() {
-      if (this.mockConfigData && this.mockConfigData.mockExpectConfigList) {
-        return this.mockConfigData.mockExpectConfigList.filter(
-          data => !this.tableSearch || data.name.toLowerCase().includes(this.tableSearch.toLowerCase()))
-      } else {
-        return [];
-      }
-    },
     redirectToTest(row) {
       let requestParam = null;
       if (row && row.request) {
