+++ frontend/src/business/components/api/definition/components/response/ResponseText.vue
@@ -56,6 +56,7 @@
         this.isActive = !this.isActive;
       },
     },
+
     mounted() {
       if (!this.response.headers) {
         return;
