+++ frontend/src/business/components/api/homepage/components/ApiInfoCard.vue
@@ -1,5 +1,5 @@
 <template>
-  <el-card class="table-card" v-loading="result.loading" body-style="padding:16px 5px 16px 16px;" style="height: 360px;margin-top: 5px">
+  <el-card class="table-card" v-loading="result.loading" body-style="padding:16px 5px 16px 16px;" style="height: 350px;margin-top: 5px">
     <div slot="header">
       <span class="title">
         {{ $t('api_test.home_page.api_count_card.title') }}
