+++ frontend/src/i18n/zh-CN.js
@@ -1265,7 +1265,6 @@ export default {
         extract_param: "提取参数",
         add_module: "创建模块",
         edit_api: "编辑接口",
-        show_api: "查看接口",
         test_plan_select: "请选择测试计划",
         create_info: '创建',
         update_info: '更新',
